pyproject.toml
[project]
name = "qarnot-mcp"
version = "0.1.0"
description = "FastMCP server for Qarnot Computing API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.5",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[project.scripts]
qarnot-mcp = "src.qarnot_mcp.server:mcp.run"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
